requirements.txt
boto3==1.3.1
botocore==1.4.24
decorator==4.0.9
Django==1.9.6
django-pylibmc==0.6.1
django-storages==1.4.1
docutils==0.12
gunicorn==19.6.0
jmespath==0.9.0
networkx==1.11
numpy==1.11.0
pkg-resources==0.0.0
progressbar2==3.6.2
pylibmc==1.5.1
pymongo==3.2.2
python-dateutil==2.5.3
scipy==0.17.1
six==1.10.0
